Created user interface for the library.
Non-automatic functions return an array of probabilities of
interest. There is no memoization, nor automatic decision of the
algorithm etc. Those functions are typically for power users or
developers who want to use the library to optimize their code.

Automatic functions also return a single number, which is the
requested probability. The results are memoized and where
applicable, an algorithm is chosen for the task. These functions
are to make the library easy to use.