Memoize various skip seed probabilities
The internal hashes are now indexed with three parameters: n (skipping),
u (divergence rate) and N (number of duplicates). This allows the
probabilities of skip seeds to be stored for various values of n. The
parameter n is not used for exact seeds and MEM seeds, so the value is
set to 0. The user can specify other values when calling `store_prob()`.

This commit fixes #1.